# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HAOS API",
    "version": "0.1.0",
    "description": "Home Assistant for Open Source - Matrix-based communication platform",
    "contact": {
      "name": "HAOS Team",
      "url": "https://github.com/aaron777collins/haos-v2"
    }
  },
  "servers": [
    {
      "url": "{protocol}://{host}/api",
      "description": "HAOS API Server",
      "variables": {
        "protocol": {
          "enum": [
            "http",
            "https"
          ],
          "default": "https"
        },
        "host": {
          "default": "localhost:3000",
          "description": "HAOS server host"
        }
      }
    }
  ],
  "paths": {
...
    "/api/members/{memberId}": {
      "patch": {
        "deprecated": true,
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "memberId identifier"
          }
        ],
        "responses": {
          "410": {
            "description": "Gone (Deprecated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "deprecated": true,
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "memberId identifier"
          }
        ],
        "responses": {
          "410": {
            "description": "Gone (Deprecated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "post": {
        "deprecated": true,
        "tags": [
          "Members"
        ],
        "parameters": [
          {
            "name": "memberId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "memberId identifier"
          }
        ],
        "responses": {
          "410": {
            "description": "Gone (Deprecated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/servers/{serverId}": {
      "patch": {
        "deprecated": true,
        "tags": [
          "Servers"
        ],
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "serverId identifier"
          }
        ],
        "responses": {
          "410": {
            "description": "Gone (Deprecated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "delete": {
        "deprecated": true,
        "tags": [
          "Servers"
        ],
        "parameters": [
          {
            "name": "serverId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "serverId identifier"
          }
        ],
        "responses": {
          "410": {
            "description": "Gone (Deprecated)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie from login endpoint"
      }
    },
...
  },
...
}